+++ modules-services2/service-templates-core/src/main/java/org/openprovenance/prov/service/dispatch/EnactorException.java
@@ -1,0 +1,6 @@
+package org.openprovenance.prov.service.dispatch;
+
+public class EnactorException extends RuntimeException {
+    public EnactorException(String s, String method) {
+    }
+}