+++ modules-services2/service-templates-core/src/main/java/org/openprovenance/prov/service/readers/TableKey.java
@@ -1,0 +1,14 @@
+package org.openprovenance.prov.service.readers;
+
+public class TableKey {
+    public String isA;
+    public Integer ID;
+
+    @Override
+    public String toString() {
+        return "TableKey{" +
+                "isA='" + isA + '\'' +
+                ", ID=" + ID +
+                '}';
+    }
+}