+++ modules-services2/service-templates-core/src/main/java/org/openprovenance/prov/service/readers/TemplatesVizConfig.java
@@ -1,0 +1,8 @@
+package org.openprovenance.prov.service.readers;
+
+public class TemplatesVizConfig {
+    public Integer id;
+    public String template;
+    public String property;
+    public String style;
+}